Deduplicate OLED ticker rendering and clear all clippy lints
- One ticker_offset with a pause_frames parameter replaces the
  pause/immediate twin implementations.
- draw_marquee extracts the two-copy wrapping-marquee drawing that was
  repeated four times (row values, both mode-view headlines, legend
  items).
- draw_mode_row extracts the duplicated MODE label row from both
  render_modify branches; push_drum_names extracts the duplicated
  comma-separated GM drum-name builder.
- Delete dead draw_segs_tickered / draw_segs_row_tickered.
- oled_render dispatches on UiMode instead of raw u8 with comments.
- Remaining clippy lints fixed crate-wide (clamp patterns, redundant
  casts, manual range-contains, iterator indexing, is_ascii_digit,
  FmtBuf Default); clippy is now clean with no warnings.

+++ src/wasm-rust/engine/src/engine_input.rs
@@ -367,7 +367,7 @@ fn pattern_press_random(s: &mut EngineState, row: i16, tick: i32, tpc: i32) {
     let is_drum = s.is_drum_channel(ch);
 
     if is_drum {
-        let pat = get_drum_pattern(row as i16, engine_random(s));
+        let pat = get_drum_pattern(row, engine_random(s));
         let amt = pat.amount;
         let speed_roll = engine_random(s) % 100;
         let space = if speed_roll < 15 && pat.space_mul >= 2 {
